ShieldEye Core Legacy/requirements.txt
python-nmap==0.7.1
requests==2.31.0
beautifulsoup4==4.12.2
reportlab==4.0.4
lxml==4.9.3
tkinter
threading
socket
subprocess
json
datetime 